mcp/agent/requirements.txt
fastmcp>=2.9.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
redis>=5.0.0
langraph>=0.1.0
langsmith>=0.1.0
aiohttp>=3.9.0
httpx>=0.27.0 